refactor: move most unit tests into separate package from code under test (black-box testing)
