+++ internal/catalog/catalog_test/testdata/freeform/some-dir/non-crd.yaml
