+++ internal/catalog/catalog_test/testdata/freeform/some-dir/prod/sub-dir/rel.yaml
