+++ internal/catalog/catalog_test/testdata/freeform/some-other-dir/dev/sub-dir/rel.yaml
