+++ internal/catalog/catalog_test/testdata/freeform/some-other-dir/dev/unknown-version-rel.yaml
