+++ internal/catalog/catalog_test/testdata/freeform/some-projects/project1.yaml
