+++ internal/catalog/catalog_test/testdata/freeform/some-projects/sub-dir/project2.yaml
